Ensure the quests are being inserted in `dict_to_quests`
Fix for #31

The issue comes from the fact JSON parses integers as floats.

The fix consists of making godot handle type conversion before  appending the array.

The solution is similar to what @fcazalet suggested, but instead the type conversion is handled by Godot itself (which makes things a bit faster)

Co-Authored-By: Florian Cazalet <707250+fcazalet@users.noreply.github.com>